requirements.txt
Flask
flask_sqlalchemy
flask-cors
python-dotenv
gunicorn
PyJWT
Werkzeug
ephem
matplotlib
gunicorn
PyJWT